Revert display layout invalidation
Fix regression made in commit 6511755 - when removing the display
layout, we also must remove the other relevant data - particulary
the character offset map. But it is all a losing game anyway, just
hash as much of the formats as we need to ensure the display layout
gets recalculated after theme change, whenever it happens.